Time: 0 ms (100.00%) | Memory: 44.2 MB (10.85%) - LeetSync

+++ 50-powx-n/powx-n.java
@@ -1,6 +1,6 @@
 class Solution {
     public double myPow(double x, int n) {
-        if(x == 0) return 0;
+        if(x == (double) 0) return 0;
         if(n == 0) return 1;
 
         double result = helper(x, Math.abs(n));
